+++ src/main/java/com/example/VivaLaTrip/Service/UserService.java
@@ -1,12 +1,12 @@
 package com.example.VivaLaTrip.Service;
 
 import com.example.VivaLaTrip.Entity.UserInfo;
+import com.example.VivaLaTrip.Form.FailureUserInfo;
 import com.example.VivaLaTrip.Repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -79,42 +79,7 @@ public class UserService implements UserDetailsService {
     public UserDetails loadUserByUsername(String id) throws UsernameNotFoundException {
         Optional<UserInfo> userInfo = userRepository.findByID(id);
         if (!userInfo.isPresent()) {
-            return new UserDetails() {
-                @Override
-                public Collection<? extends GrantedAuthority> getAuthorities() {
-                    return null;
-                }
-
-                @Override
-                public String getPassword() {
-                    return null;
-                }
-
-                @Override
-                public String getUsername() {
-                    return null;
-                }
-
-                @Override
-                public boolean isAccountNonExpired() {
-                    return false;
-                }
-
-                @Override
-                public boolean isAccountNonLocked() {
-                    return false;
-                }
-
-                @Override
-                public boolean isCredentialsNonExpired() {
-                    return false;
-                }
-
-                @Override
-                public boolean isEnabled() {
-                    return false;
-                }
-            };
+            return new FailureUserInfo();
         }
         return new User(userInfo.get().getID(), userInfo.get().getPassword(), userInfo.get().getAuthorities());
     }
